Modularized the code + tested
