+++ src/experiments.py
@@ -1,834 +1,10 @@
-import json
-import time
 import os
-import openai
-from openai import OpenAI
-import time
+import json
 import pickle
-from rouge_score import rouge_scorer
-import numpy as np
 from tqdm import tqdm
-from statistics import mode
+from openai import OpenAI
 
-class LLMAdapter:
-    def __init__(self):
-        # CWEs IDs and names
-        self.cwes = {
-            "cwe-22": "path traversal",
-            "cwe-77": "OS command injection",
-            "cwe-79": "cross-site scripting",
-            "cwe-89": "SQL injection",
-            "cwe-190": "integer overflow",
-            "cwe-416": "use after free",
-            "cwe-476": "NULL pointer dereference",
-            "cwe-787": "out-of-bounds write"
-        }
-        # CWEs definitions from MITRE website
-        self.defs = json.loads(open("datasets/mitre-cwe-definitions.json", "r", encoding='utf-8').read())
-        # CWEs languages
-        self.lang = {
-            "cwe-22": "c",
-            "cwe-77": "c",
-            "cwe-79": "py",
-            "cwe-89": "py",
-            "cwe-190": "c",
-            "cwe-416": "c",
-            "cwe-476": "c",
-            "cwe-787": "c"
-        }
-        # Function map for Prompts
-        self.prompts = {
-            "promptS1": self.promptS1,
-            "promptS2": self.promptS2,
-            "promptS3": self.promptS3,
-            "promptS4": self.promptS4,
-            "promptS5": self.promptS5,
-            "promptS6": self.promptS6,
-            "promptR1": self.promptR1,
-            "promptR2": self.promptR2,
-            "promptR3": self.promptR3,
-            "promptR4": self.promptR4,
-            "promptR5": self.promptR5,
-            "promptR6": self.promptR6,
-            "promptD1": self.promptD1,
-            "promptD2": self.promptD2,
-            "promptD3": self.promptD3,
-            "promptD4": self.promptD4,
-            "promptD5": self.promptD5
-        }
-        # Map for Prompts types
-        self.prompts_map = {
-            'promptS1': ('ZS', 'TO'),
-            'promptS2': ('ZS', 'RO'),
-            'promptS3': ('ZS', 'RO'),
-            'promptS4': ('ZS', 'RO'),
-            'promptS5': ('FS', 'TO'),
-            'promptS6': ('FS', 'RO'),
-            'promptR1': ('ZS', 'TO'),
-            'promptR2': ('ZS', 'RO'),
-            'promptR3': ('ZS', 'TO'),
-            'promptR4': ('FS', 'RO'),
-            'promptR5': ('FS', 'RO'),
-            'promptR6': ('FS', 'TO'),
-            'promptD1': ('ZS', 'TO'),
-            'promptD2': ('ZS', 'RO'),
-            'promptD3': ('FS', 'RO'),
-            'promptD4': ('FS', 'RO'),
-            'promptD5': ('FS', 'TO'),
-        }
-
-    #################################
-    # Adapter Setup Functions 
-    # (To be overwriten)
-    #################################
-
-    # Pass named arguments in prepare() to initialize the adapter
-    def prepare(self, **kwargs):
-        pass
-
-    # Prepare the prompt
-    def prepare_prompt(self, **kwargs):
-        pass
-
-    # Chat with LLM and return the response
-    def chat(self, **kwargs):
-        pass
-
-    ############################
-    # Standard Prompts
-    ############################
-
-    # Statndard Prompt 1 (ZS) (TO)
-    def promptS1(self, **kwargs):
-        cwe = kwargs['cwe']
-        cwe_name = self.cwes[cwe]
-        code = kwargs['code']
-
-        # system
-        sys = ""
-
-        # example
-        exp = []
-
-        # text
-        q = "Does this code contain instances of the security vulnerability known as {}?".format(cwe_name)
-        text = self.prepare_prompt(code=code, question=q, prompt="Question")
-
-        # response
-        response = self.chat(system=sys, examples=exp, text=text)
-
-        return response
-    
-    # Statndard Prompt 2 (ZS) (RO)
-    def promptS2(self, **kwargs):
-        cwe = kwargs['cwe']
-        cwe_name = self.cwes[cwe]
-        code = kwargs['code']
-        
-        # system
-        sys = "You are a helpful assistant."
-
-        # example
-        exp = []
-
-        # text
-        q = "Does this code contain instances of the security vulnerability known as {}?".format(cwe_name)
-        text = self.prepare_prompt(code=code, question=q, prompt="Question")
-
-        # response
-        response = self.chat(system=sys, examples=exp, text=text)
-
-        return response
-    
-    # Statndard Prompt 3 (ZS) (RO)
-    def promptS3(self, **kwargs):
-        cwe = kwargs['cwe']
-        cwe_name = self.cwes[cwe]
-        code = kwargs['code']
-        
-        # system
-        sys = "You are a code security expert."
-
-        # example
-        exp = []
-
-        # text
-        q = "Does this code contain instances of the security vulnerability known as {}?".format(cwe_name)
-        text = self.prepare_prompt(code=code, question=q, prompt="Question")
-
-        # response
-        response = self.chat(system=sys, examples=exp, text=text)
-
-        return response
-    
-    # Statndard Prompt 4 (ZS) (RO)
-    def promptS4(self, **kwargs):
-        cwe = kwargs['cwe']
-        cwe_name = self.cwes[cwe]
-        code = kwargs['code']
-        
-        # system
-        sys = "You are a code security expert who analyzes the given code for the security vulnerability known as {}.".format(cwe_name)
-
-        # example
-        exp = []
-
-        # text
-        text = self.prepare_prompt(code=code, prompt="Analysis")
-
-        # response
-        response = self.chat(system=sys, examples=exp, text=text)
-
-        return response
-
-    # Statndard Prompt 5 (FS) (TO)
-    def promptS5(self, **kwargs):
-        cwe = kwargs['cwe']
-        cwe_name = self.cwes[cwe]
-        
-        # system
-        sys = ""
-
-        # example
-        q = "Does this code contain instances of the security vulnerability known as {}?".format(cwe_name)
-        vuln_code = open("datasets/few-shot/examples/{}.{}".format(cwe, self.lang[cwe]), "r", encoding="utf-8").read()
-        patch_code = open("datasets/few-shot/examples/{}-p.{}".format(cwe, self.lang[cwe]), "r", encoding="utf-8").read()
-        vuln_reason = open("datasets/few-shot/reasoning/{}.txt".format(cwe), "r",encoding="utf-8").read()
-        patch_reason = open("datasets/few-shot/reasoning/{}-p.txt".format(cwe), "r", encoding="utf-8").read()
-        vuln_prompt = self.prepare_prompt(code=vuln_code, question=q, prompt="Question")
-        patch_prompt = self.prepare_prompt(code=patch_code, question=q, prompt="Question")
-        vuln = (vuln_prompt, vuln_reason)
-        patch = (patch_prompt, patch_reason)
-        exp = [vuln, patch]
-
-        # text
-        prompt = self.prepare_prompt(code=kwargs['code'], question=q, prompt="Question")
-        text = prompt
-
-        # response
-        response = self.chat(system=sys, examples=exp, text=text)
-
-        return response
-    
-    # Statndard Prompt 6 (FS) (RO)
-    def promptS6(self, **kwargs):
-        cwe = kwargs['cwe']
-        cwe_name = self.cwes[cwe]
-        
-        # system
-        sys = "You are a code security expert who analyzes the given code for the security vulnerability known as {}.".format(cwe_name)
-
-        # example
-        vuln_code = open("datasets/few-shot/examples/{}.{}".format(cwe, self.lang[cwe]), "r", encoding="utf-8").read()
-        patch_code = open("datasets/few-shot/examples/{}-p.{}".format(cwe, self.lang[cwe]), "r", encoding="utf-8").read()
-        vuln_reason = open("datasets/few-shot/reasoning/{}.txt".format(cwe), "r", encoding="utf-8").read()
-        patch_reason = open("datasets/few-shot/reasoning/{}-p.txt".format(cwe), "r", encoding="utf-8").read()
-        vuln_prompt = self.prepare_prompt(code=vuln_code, prompt="Analysis")
-        patch_prompt = self.prepare_prompt(code=patch_code, prompt="Analysis")
-        vuln = (vuln_prompt, vuln_reason)
-        patch = (patch_prompt, patch_reason)
-        exp = [vuln, patch]
-
-        # text
-        text = self.prepare_prompt(code=kwargs['code'], prompt="Analysis")
-
-        # response
-        response = self.chat(system=sys, examples=exp, text=text)
-
-        return response
-
-    ############################
-    # Step-by-Step Prompts
-    ############################
-
-    # Step-by-Step Prompt 1 (ZS) (TO)
-    def promptR1(self, **kwargs):
-        cwe = kwargs['cwe']
-        cwe_name = self.cwes[cwe]
-        code = kwargs['code']
-
-        # system
-        sys = ""
-
-        # example
-        exp = []
-
-        # text
-        q = "Does this code contain instances of the security vulnerability known as {}?".format(cwe_name)
-        answer = "Let's think step by step."
-        prompt = self.prepare_prompt(code=code, question=q, answer=answer, prompt="Q/A")
-        text = prompt
-
-        # response
-        response = self.chat(system=sys, examples=exp, text=text)
-
-        return response
-    
-    # Step-by-Step Prompt 2 (ZS) (RO)
-    def promptR2(self, **kwargs):
-        cwe = kwargs['cwe']
-        cwe_name = self.cwes[cwe]
-        code = kwargs['code']
-
-        # system
-        sys = "You are a code security expert who analyzes the given code for the security vulnerability known as {} following these four steps:\n1. First you describe the overview of the code\n2. Then based on the overview you identify the sub-components in code that could lead to {}\n3. After that you do a detailed analysis of the identified sub-components for the existence of the {}\n4. Based on the detailed analysis you decide and answer whether the {} is present in the given code or not".format(cwe_name, cwe_name, cwe_name + " vulnerability", cwe_name + " vulnerability")
-
-        # example
-        exp = []
-
-        # text
-        text = self.prepare_prompt(code=code, prompt="Analysis")
-
-        # response
-        response = self.chat(system=sys, examples=exp, text=text)
-
-        return response
-    
-    # Step-by-Step Prompt 3 (ZS) (TO)
-    def promptR3(self, **kwargs):
-        cwe = kwargs['cwe']
-        cwe_name = self.cwes[cwe]
-        code = kwargs['code']
-
-        # system
-        sys = ""
-
-        # example
-        exp = []
-
-        # overview
-        prompt_code = self.prepare_prompt(code=code, prompt="None")
-        overview = "Provide a brief overview of the code."
-        prompt = prompt_code + "\n\n" + overview
-        overview_response = self.chat(system=sys, examples=exp, text=prompt)
-        exp.append((prompt, overview_response))
-
-        # sub-components
-        sub_components = "Based on the overview identify the sub-components in code that could lead to a security vulnerability knows as {}.".format(cwe_name)
-        sub_response = self.chat(system=sys, examples=exp, text=sub_components)
-        exp.append((sub_components, sub_response))
-
-        # detailed analysis
-        detailed_analysis = "Perform a detailed analysis of the identified sub-components for the existence of the {}.".format(cwe_name + " vulnerability")
-        analysis_response = self.chat(system=sys, examples=exp, text=detailed_analysis)
-        exp.append((detailed_analysis, analysis_response))
-
-        # answer
-        answer = "Based on the detailed analysis decide and answer whether the {} is present in the given code or not.".format(cwe_name + " vulnerability")
-        ans_response = self.chat(system=sys, examples=exp, text=answer)
-
-        # response
-        response = "Overview:\n" + overview_response + "\n\nSub-cmponents:\n" + sub_response + "\n\nDetailed Analysis:\n" + analysis_response + "\n\nAnswer:\n" + ans_response
-        return response
-
-    # Step-by-Step Prompt 4 (FS) (RO)
-    def promptR4(self, **kwargs):
-        cwe = kwargs['cwe']
-        cwe_name = self.cwes[cwe]
-
-        # system
-        sys = "You are a code security expert who analyzes the given code for the security vulnerability known as {}.".format(cwe_name)
-
-        # example
-        vuln_code = open("datasets/few-shot/examples/{}.{}".format(cwe, self.lang[cwe]), "r", encoding="utf-8").read()
-        patch_code = open("datasets/few-shot/examples/{}-p.{}".format(cwe, self.lang[cwe]), "r", encoding="utf-8").read()
-        vuln_reason = open("datasets/few-shot/step-by-step-reasoning/{}.txt".format(cwe), "r", encoding="utf-8").read()
-        patch_reason = open("datasets/few-shot/step-by-step-reasoning/{}-p.txt".format(cwe), "r", encoding="utf-8").read()
-        vuln_prompt = self.prepare_prompt(code=vuln_code, prompt="Analysis")
-        patch_prompt = self.prepare_prompt(code=patch_code, prompt="Analysis")
-        vuln = (vuln_prompt, vuln_reason)
-        patch = (patch_prompt, patch_reason)
-        exp = [vuln, patch]
-
-        # text
-        text = self.prepare_prompt(code=kwargs['code'], prompt="Analysis")
-
-        # response
-        response = self.chat(system=sys, examples=exp, text=text)
-
-        return response
-
-    # Step-by-Step Prompt 5 (FS) (RO)
-    def promptR5(self, **kwargs):
-        cwe = kwargs['cwe']
-        cwe_name = self.cwes[cwe]
-
-        # system
-        sys = "You are a code security expert who analyzes the given code for the security vulnerability known as {} following these four steps:\n1. First you describe the overview of the code\n2. Then based on the overview you identify the sub-components in code that could lead to {}\n3. After that you do a detailed analysis of the identified sub-components for the existence of the {}\n4. Based on the detailed analysis you decide and answer whether the {} is present in the given code or not".format(cwe_name, cwe_name, cwe_name + " vulnerability", cwe_name + " vulnerability")
-
-        # example
-        vuln_code = open("datasets/few-shot/examples/{}.{}".format(cwe, self.lang[cwe]), "r", encoding="utf-8").read()
-        patch_code = open("datasets/few-shot/examples/{}-p.{}".format(cwe, self.lang[cwe]), "r", encoding="utf-8").read()
-        vuln_reason = open("datasets/few-shot/step-by-step-reasoning/{}.txt".format(cwe), "r", encoding="utf-8").read()
-        patch_reason = open("datasets/few-shot/step-by-step-reasoning/{}-p.txt".format(cwe), "r", encoding="utf-8").read()
-        vuln_prompt = self.prepare_prompt(code=vuln_code, prompt="Analysis")
-        patch_prompt = self.prepare_prompt(code=patch_code, prompt="Analysis")
-        vuln = (vuln_prompt, vuln_reason)
-        patch = (patch_prompt, patch_reason)
-        exp = [vuln, patch]
-
-        # text
-        text = self.prepare_prompt(code=kwargs['code'], prompt="Analysis")
-
-        # response
-        response = self.chat(system=sys, examples=exp, text=text)
-
-        return response
-    
-    # Step-by-Step Prompt 6 (FS) (TO)
-    def promptR6(self, **kwargs):
-        cwe = kwargs['cwe']
-        cwe_name = self.cwes[cwe]
-
-        # system
-        sys = "Guidelines to analyze the given code for the security vulnerability known as {}:\n1. First describe the overview of the code\n2. Then based on the overview identify the sub-components in code that could lead to {}\n3. After that do a detailed analysis of the identified sub-components for the existence of the {}\n4. Based on the detailed analysis decide and answer whether the {} is present in the given code or not".format(cwe_name, cwe_name, cwe_name + " vulnerability", cwe_name + " vulnerability")
-
-        # example
-        vuln_code = open("datasets/few-shot/examples/{}.{}".format(cwe, self.lang[cwe]), "r", encoding="utf-8").read()
-        patch_code = open("datasets/few-shot/examples/{}-p.{}".format(cwe, self.lang[cwe]), "r", encoding="utf-8").read()
-        vuln_reason = open("datasets/few-shot/step-by-step-reasoning/{}.txt".format(cwe), "r", encoding="utf-8").read()
-        patch_reason = open("datasets/few-shot/step-by-step-reasoning/{}-p.txt".format(cwe), "r", encoding="utf-8").read()
-        vuln_prompt = self.prepare_prompt(code=vuln_code, prompt="Analysis")
-        patch_prompt = self.prepare_prompt(code=patch_code, prompt="Analysis")
-        vuln = (vuln_prompt, vuln_reason)
-        patch = (patch_prompt, patch_reason)
-        exp = [vuln, patch]
-
-        # text
-        text = self.prepare_prompt(code=kwargs['code'], prompt="Analysis")
-
-        # response
-        response = self.chat(system=sys, examples=exp, text=text)
-
-        return response
-
-    ############################
-    # Definition Prompts
-    ############################
-
-    # Definition Prompt 1 (ZS) (TO)
-    def promptD1(self, **kwargs):
-        cwe = kwargs['cwe']
-        cwe_name = self.cwes[cwe]
-        code = kwargs['code']
-
-        # system
-        sys = ""
-
-        # example
-        exp = []
-
-        # text
-        q = "Does this code contain instances of the security vulnerability known as {}?".format(cwe_name)
-        text = self.prepare_prompt(defin=self.defs[cwe], code=code, question=q, prompt="Def")
-
-        # response
-        response = self.chat(system=sys, examples=exp, text=text)
-
-        return response
-
-    # Definition Prompt 2 (ZS) (RO)
-    def promptD2(self, **kwargs):
-        cwe = kwargs['cwe']
-        cwe_name = self.cwes[cwe]
-        code = kwargs['code']
-
-        # system
-        sys = "You are a code security expert who analyzes the given code for the security vulnerability known as {}.\n\n{}".format(cwe_name, self.defs[cwe])
-
-        # example
-        exp = []
-
-        # text
-        text = self.prepare_prompt(code=code, prompt="Analysis")
-
-        # response
-        response = self.chat(system=sys, examples=exp, text=text)
-
-        return response
-
-    # Definition Prompt 3 (FS) (RO)
-    def promptD3(self, **kwargs):
-        cwe = kwargs['cwe']
-        cwe_name = self.cwes[cwe]
-
-        # system
-        sys = "You are a code security expert who analyzes the given code for the security vulnerability known as {}.\n\n{}".format(cwe_name, self.defs[cwe])
-
-        # example
-        vuln_code = open("datasets/few-shot/examples/{}.{}".format(cwe, self.lang[cwe]), "r", encoding="utf-8").read()
-        patch_code = open("datasets/few-shot/examples/{}-p.{}".format(cwe, self.lang[cwe]), "r", encoding="utf-8").read()
-        vuln_reason = open("datasets/few-shot/reasoning/{}.txt".format(cwe), "r", encoding="utf-8").read()
-        patch_reason = open("datasets/few-shot/reasoning/{}-p.txt".format(cwe), "r", encoding="utf-8").read()
-        vuln_prompt = self.prepare_prompt(code=vuln_code, prompt="Analysis")
-        patch_prompt = self.prepare_prompt(code=patch_code, prompt="Analysis")
-        vuln = (vuln_prompt, vuln_reason)
-        patch = (patch_prompt, patch_reason)
-        exp = [vuln, patch]
-
-        # text
-        text = self.prepare_prompt(code=kwargs['code'], prompt="Analysis")
-
-        # response
-        response = self.chat(system=sys, examples=exp, text=text)
-
-        return response
-
-    # Definition Prompt 4 (FS) (RO)
-    def promptD4(self, **kwargs):
-        cwe = kwargs['cwe']
-        cwe_name = self.cwes[cwe]
-
-        # system
-        sys = "You are a code security expert who analyzes the given code for the security vulnerability known as {}.\n\n{}".format(cwe_name, self.defs[cwe])
-
-        # example
-        vuln_code = open("datasets/few-shot/examples/{}.{}".format(cwe, self.lang[cwe]), "r", encoding="utf-8").read()
-        patch_code = open("datasets/few-shot/examples/{}-p.{}".format(cwe, self.lang[cwe]), "r", encoding="utf-8").read()
-        vuln_reason = open("datasets/few-shot/step-by-step-reasoning/{}.txt".format(cwe), "r", encoding="utf-8").read()
-        patch_reason = open("datasets/few-shot/step-by-step-reasoning/{}-p.txt".format(cwe), "r", encoding="utf-8").read()
-        vuln_prompt = self.prepare_prompt(code=vuln_code, prompt="Analysis")
-        patch_prompt = self.prepare_prompt(code=patch_code, prompt="Analysis")
-        vuln = (vuln_prompt, vuln_reason)
-        patch = (patch_prompt, patch_reason)
-        exp = [vuln, patch]
-
-        # text
-        text = self.prepare_prompt(code=kwargs['code'], prompt="Analysis")
-
-        # response
-        response = self.chat(system=sys, examples=exp, text=text)
-
-        return response
-    
-    # Definition Prompt 5 (FS) (TO)
-    def promptD5(self, **kwargs):
-        cwe = kwargs['cwe']
-        cwe_name = self.cwes[cwe]
-
-        # system
-        sys = "Analyze the given code for the security vulnerability known as {}.\n\n{}".format(cwe_name, self.defs[cwe])
-
-        # example
-        vuln_code = open("datasets/few-shot/examples/{}.{}".format(cwe, self.lang[cwe]), "r", encoding="utf-8").read()
-        patch_code = open("datasets/few-shot/examples/{}-p.{}".format(cwe, self.lang[cwe]), "r", encoding="utf-8").read()
-        vuln_reason = open("datasets/few-shot/step-by-step-reasoning/{}.txt".format(cwe), "r", encoding="utf-8").read()
-        patch_reason = open("datasets/few-shot/step-by-step-reasoning/{}-p.txt".format(cwe), "r", encoding="utf-8").read()
-        vuln_prompt = self.prepare_prompt(code=vuln_code, prompt="Analysis")
-        patch_prompt = self.prepare_prompt(code=patch_code, prompt="Analysis")
-        vuln = (vuln_prompt, vuln_reason)
-        patch = (patch_prompt, patch_reason)
-        exp = [vuln, patch]
-
-        # text
-        text = self.prepare_prompt(code=kwargs['code'], prompt="Analysis")
-
-        # response
-        response = self.chat(system=sys, examples=exp, text=text)
-
-        return response
-
-    ############################
-    # Helper Functions
-    ############################
-
-    def fix_edge_cases(self, text):
-        # Llama long same letter word 'string' edge case
-        max_length = 100
-        len_list = list(map(len, text.split()))
-        if len(len_list) and max(len_list) > max_length:
-            words = text.split()
-            truncated_words = [word[:max_length] if len(word) > max_length else word for word in words]
-            text = ' '.join(truncated_words)
-            return True, text
-        return False, text
-
-    def extract_pred(self, **kwargs):
-        cwe = kwargs['cwe']
-        content = kwargs['text']
-        q = "As a final decision or answer, does the text state that the code contains a security vulnerability known as {}? (Answer yes or no or n/a, incase the text does not provide a definite answer)".format(self.cwes[cwe])
-        text = 'Text: """\n' + content + '\n"""\n' + q
-
-        done, itr = False, 5
-        while not done and itr:
-            try:
-                # Applied fix
-                fix_required = self.fix_edge_cases(text)
-                if fix_required[0]:
-                    text = fix_required[1]
-                gpt_response = self.gpt_client.chat.completions.create(
-                    model = "gpt-4",
-                    temperature = 0.0,
-                    messages = [{"role": "user", "content": text}]
-                )
-                done = True
-            except openai.RateLimitError as e:
-                print("Rate limit reached. Waiting for 60 seconds...")
-                time.sleep(60)
-            except openai.APIError as e:
-                print(f"Error in extract_pred {e}")
-                # Applied fix
-                fix_required = self.fix_edge_cases(text)
-                if fix_required[0]:
-                    text = fix_required[1]
-                itr -= 1
-                print("Retrying...")
-        if done:
-            response = gpt_response.choices[0].message.content
-            pred = response.lower()
-            return pred
-        return None
-
-    def extract_info(self, **kwargs):
-        cwe = kwargs['cwe']
-        sys = "You are a helpful assistant who extracts answer whether an instance of vulnerability is present or not and its reason 'why' from the given text in the following format:\nAnswer: ...\nReason: ...\n\nYou must follow these rules while extracting information:\n1. As a final decision or answer, does the text state that the code contains a security vulnerability known as {}? (Answer yes or no or n/a, incase the text does not provide a definite answer)\n2. If the text does not state the reason 'why' an instance of vulnerability is present or not then just write 'n/a' in reason.\n3. If the answer in 1 is 'n/a' then just write 'n/a' in reason.\n4. Only if 2 and 3 are not true then briefly describe the reasons mentioned in the text that state 'why' the code does or does not contain a security vulnerability known as {}. Max word limit for reason is 100 words. Write in terms of code, e.g., 'The code/program/function ...'".format(self.cwes[cwe], self.cwes[cwe])
-        content = kwargs["text"]
-
-        done, itr = False, 5
-        while not done and itr:
-            try:
-                # Applied fix
-                fix_required = self.fix_edge_cases(content)
-                if fix_required[0]:
-                    content = fix_required[1]
-                gpt_response = self.gpt_client.chat.completions.create(
-                    model = "gpt-4",
-                    temperature = 0.0,
-                    messages = [{"role": "system", "content": sys}, {"role": "user", "content": content}]
-                )
-                done = True
-            except openai.RateLimitError as e:
-                print("Rate limit reached. Waiting for 60 seconds...")
-                time.sleep(60)
-            except openai.APIError as e:
-                print(f"Error in extract_info {e}")
-                # Applied fix
-                fix_required = self.fix_edge_cases(content)
-                if fix_required[0]:
-                    content = fix_required[1]
-                itr -= 1
-                print("Retrying...")
-        
-        pred, reason = None, None
-        if done:
-            response = gpt_response.choices[0].message.content
-            pred = response.split('\n')[0].split('Answer:')[-1][1:].lower()
-            if pred == "n/a":
-                reason = "n/a"
-            else:
-                reason = response.split('Reason:')[-1][1:]
-        return pred, reason
-
-    def rouge(self, **kwargs):
-        scorer = rouge_scorer.RougeScorer(['rouge1', 'rougeL'], use_stemmer=True)
-        scores = scorer.score(kwargs['reason'], kwargs['ground_truth'])
-        return scores['rouge1'].precision
-
-    def cos_similarity(self, **kwargs):
-        done, itr = False, 5
-        while not done and itr:
-            try:
-                resp = self.gpt_client.embeddings.create(
-                    input=[kwargs['reason']],
-                    model="text-embedding-3-large"
-                )
-
-                embedding_a = resp.data[0].embedding
-                done = True
-            except openai.RateLimitError as e:
-                print("Rate limit reached. Waiting for 60 seconds...")
-                time.sleep(60)
-            except openai.APIError as e:
-                print(f"Error in cos_similarity {e}")
-                itr -= 1
-                print("Retrying...")
-        
-        similarity_score = None
-        if done:
-            similarity_score = np.dot(embedding_a, kwargs['ground_truth'])
-        return similarity_score
-
-    def gpt_eval(self, **kwargs):
-        reason = kwargs['reason']
-        gt = kwargs['ground_truth']
-
-        q = "Evaluate if the reasoning in the response aligns with the ground truth. (Answer yes or no)"
-        prompt = q + '\n\nResponse: """\n' + reason + '\n"""\n\nGround Truth: """\n' + gt + '\n"""'
-
-        itr = 5
-        while itr:
-            try:
-                gpt_response = self.gpt_client.chat.completions.create(
-                    model = "gpt-4",
-                    temperature = 0.0,
-                    messages = [{"role": "user", "content": prompt}]
-                )
-                answer = gpt_response.choices[0].message.content.lower()
-                return answer
-            except openai.RateLimitError as e:
-                print("Rate limit reached. Waiting for 60 seconds...")
-                time.sleep(60)
-            except openai.APIError as e:
-                print(f"Error in gpt_eval {e}")
-                itr -= 1
-                print("Retrying...")
-        return None
-
-    def check_consistency(self, **kwargs):
-        prompt = kwargs['prompt']
-        temp = kwargs['temp']
-        model = kwargs['model']
-        result_path = kwargs['result_path']
-        result_name = 'temp'
-        result_full_path = os.path.join(result_path, result_name + ".json")
-
-        # Load results
-        results = json.loads(open(result_full_path, "r", encoding='utf-8').read())
-
-        for cwe in results[prompt][temp][model]:
-            for file in results[prompt][temp][model][cwe]:
-                first_pred = results[prompt][temp][model][cwe][file]["1"]["pred"]
-                for i in results[prompt][temp][model][cwe][file]:
-                    pred = results[prompt][temp][model][cwe][file][i]["pred"]
-                    if first_pred != pred:
-                        return True
-        return False
-
-    def handle_reason(self, r, c, g):
-        ROUGE_THRES = 0.34
-        COS_SIM_THRES = 0.84
-        return mode([1 if r >= ROUGE_THRES else 0, 1 if c >= COS_SIM_THRES else 0, 1 if g=='yes' else 0])
-
-    def get_score(self, data):
-        W1, W2, W3 = 0.33, 0.33, 0.33
-        res_rate = data['total_answered']['val'] / (data['total_answered']['val'] + data['no_answer']['val'])
-        acc_rate = data['correct']['val'] / data['total_answered']['val']
-        rea_rate = data['correct_pred_correct_reason']['val'] / (data['correct_pred_correct_reason']['val'] + data['correct_pred_incorrect_reason']['val'])
-        return (W1 * res_rate) + (W2 * acc_rate) + (W3 * rea_rate)
-
-    def get_best_prompt(self, data, model, prompts):
-        max_score = 0
-        max_prompt = ''
-        for p in prompts:
-            score = self.get_score(data[model][p])
-            if score > max_score:
-                max_score = score
-                max_prompt = p
-        return max_score, max_prompt
-
-    def get_model_best_prompts(self, data, model, zt, zr, ft, fr):
-        # print("Model:", model)
-        zt_s, zt_p = self.get_best_prompt(data, model, zt)
-        zr_s, zr_p = self.get_best_prompt(data, model, zr)
-        ft_s, ft_p = self.get_best_prompt(data, model, ft)
-        fr_s, fr_p = self.get_best_prompt(data, model, fr)
-        
-        zs = zt_p if zt_s > zr_s else zr_p
-        fs = ft_p if ft_s > fr_s else fr_p
-
-        result = {
-            'ZT': zt_p,
-            'ZR': zr_p,
-            'ZS': zs,
-            'FT': ft_p,
-            'FR': fr_p,
-            'FS': fs
-        }
-        return result
-
-    def find_best_prompts(self, **kwargs):
-        model = kwargs['model']
-        result_path = kwargs['result_path']
-
-        # Load results
-        result_name = model + '.json'
-        result_full_path = os.path.join(result_path, result_name)
-        results = json.loads(open(result_full_path, "r", encoding='utf-8').read())
-
-        # metric_data
-        metric_data = {}
-        # -> model
-        if model not in metric_data:
-            metric_data[model] = {}
-        for cwe in results:
-            for file in results[cwe]:
-                for prompt in results[cwe][file]:
-                    # -> prompt
-                    if prompt not in metric_data[model]:
-                        metric_data[model][prompt] = {
-                            "total_answered": {'val': 0, 'id': []},
-                            "no_answer": {'val': 0, 'id': []},
-                            "correct": {'val': 0, 'id': []},
-                            "total_reasoned": {'val': 0, 'id': []},
-                            "no_reason": {'val': 0, 'id': []},
-                            "correct_pred_correct_reason": {'val': 0, 'id': []},
-                            "incorrect_pred_correct_reason": {'val': 0, 'id': []},
-                            "correct_pred_incorrect_reason": {'val': 0, 'id': []},
-                            "incorrect_pred_incorrect_reason": {'val': 0, 'id': []}
-                        }
-                    pred = results[cwe][file][prompt]['pred']
-                    # not answered
-                    if pred != 'yes' and pred != 'no':
-                        metric_data[model][prompt]['no_answer']['val'] += 1
-                        metric_data[model][prompt]['no_answer']['id'].append((cwe, file, prompt))
-                    # total answered
-                    else:
-                        metric_data[model][prompt]['total_answered']['val'] += 1
-                        
-                        # correct
-                        label = 0 if file[0] == 'p' else 1
-                        p = 1 if pred == 'yes' else 0
-                        correct = False
-                        if p == label:
-                            correct = True
-                            metric_data[model][prompt]['correct']['val'] += 1
-                            metric_data[model][prompt]['correct']['id'].append((cwe, file, prompt))
-                        
-                        # reason
-                        reason = results[cwe][file][prompt]['reason']
-                        if reason == 'n/a':
-                            metric_data[model][prompt]['no_reason']['val'] += 1
-                            metric_data[model][prompt]['no_reason']['id'].append((cwe, file, prompt))
-                        else:
-                            metric_data[model][prompt]['total_reasoned']['val'] += 1
-                            
-                            rouge = results[cwe][file][prompt]['rouge']
-                            cos = results[cwe][file][prompt]['cos_sim']
-                            gpt = results[cwe][file][prompt]['gpt_eval']
-                            
-                            # correct reason
-                            if self.handle_reason(rouge, cos, gpt) == 1:
-                                if correct:
-                                    metric_data[model][prompt]['correct_pred_correct_reason']['val'] += 1
-                                    metric_data[model][prompt]['correct_pred_correct_reason']['id'].append((cwe, file, prompt))
-                                else:
-                                    metric_data[model][prompt]['incorrect_pred_correct_reason']['val'] += 1
-                                    metric_data[model][prompt]['incorrect_pred_correct_reason']['id'].append((cwe, file, prompt))
-                            #incorrect reason
-                            else:
-                                if correct:
-                                    metric_data[model][prompt]['correct_pred_incorrect_reason']['val'] += 1
-                                    metric_data[model][prompt]['correct_pred_incorrect_reason']['id'].append((cwe, file, prompt))
-                                else:
-                                    metric_data[model][prompt]['incorrect_pred_incorrect_reason']['val'] += 1
-                                    metric_data[model][prompt]['incorrect_pred_incorrect_reason']['id'].append((cwe, file, prompt))
-        
-        # get best prompts
-        zs_to, zs_ro, fs_to, fs_ro = [], [], [], []
-
-        for p in self.prompts_map:
-            if self.prompts_map[p][0] == 'ZS' and self.prompts_map[p][1] == 'TO':
-                zs_to.append(p)
-            elif self.prompts_map[p][0] == 'ZS' and self.prompts_map[p][1] == 'RO':
-                zs_ro.append(p)
-            elif self.prompts_map[p][0] == 'FS' and self.prompts_map[p][1] == 'TO':
-                fs_to.append(p)
-            elif self.prompts_map[p][0] == 'FS' and self.prompts_map[p][1] == 'RO':
-                fs_ro.append(p)
-
-        best_prompts_json = self.get_model_best_prompts(metric_data, model, zs_to, zs_ro, fs_to, fs_ro)
-        return best_prompts_json
-
+class ExperimentRunner:
     ############################
     # Experiments
     ############################
@@ -1241,7 +417,8 @@ class LLMAdapter:
                             if cwe not in results[scenario]['A0']:
                                 results[scenario]['A0'][cwe] = {}
                             if file not in results[scenario]['A0'][cwe]:
-                                prompts_data = json.loads(open(os.path.join('results/prompts', model + '.json'), "r", encoding='utf-8').read())
+                                prompts_path = os.path.join(os.path.join(os.path.dirname(__file__)), "..", "results", "prompts", model + ".json")
+                                prompts_data = json.loads(open(prompts_path, "r", encoding='utf-8').read())
                                 results[scenario]['A0'][cwe][file] = prompts_data[cwe][file][prompt]
 
                             label = 0 if file[0] == 'p' else 1
@@ -1693,11 +870,11 @@ class LLMAdapter:
         '''
         api_key = kwargs['api_key']
         model = kwargs['model']
-        dataset_path = 'datasets'
+        dataset_path = os.path.join(os.path.dirname(__file__), '..', 'datasets')
 
         # make `results` directory
-        result_path = 'results'
-        os.system('mkdir ' + result_path)
+        result_path = os.path.join(os.path.dirname(__file__), '..', 'results')
+        os.makedirs(result_path, exist_ok=True)
 
         ### Deterministic Responses Experiment!!
         print("\n#######################################")
@@ -1706,7 +883,7 @@ class LLMAdapter:
 
         # Create directory to store results
         det_res_result_path = os.path.join(result_path, 'determinism')
-        os.system('mkdir ' + det_res_result_path)
+        os.makedirs(det_res_result_path, exist_ok=True)
         # Files selection for the experiment
         cwe_files = [("cwe-787", "2.c"), ("cwe-787", "p_2.c"), ("cwe-89", "2.py"), ("cwe-89", "p_2.py")]
         # Run on selected prompts
@@ -1736,7 +913,7 @@ class LLMAdapter:
 
         # Create directory to store results
         range_param_result_path = os.path.join(result_path, 'range-params')
-        os.system('mkdir ' + range_param_result_path)
+        os.makedirs(range_param_result_path, exist_ok=True)
         # Files, prompt, and temperature range selection for the experiment
         cwe_files = [("cwe-787", "3.c"), ("cwe-787", "p_3.c"), ("cwe-89", "3.py"), ("cwe-89", "p_3.py")]
         prompt = 'promptS4'
@@ -1765,7 +942,7 @@ class LLMAdapter:
 
         # Create directory to store results
         prompts_result_path = os.path.join(result_path, 'prompts')
-        os.system('mkdir ' + prompts_result_path)
+        os.makedirs(prompts_result_path, exist_ok=True)
         # Run on all prompts
         self.run_prompts_experiments(
             api_key=api_key,
@@ -1794,7 +971,7 @@ class LLMAdapter:
 
         # Create directory to store results
         robustness_result_path = os.path.join(result_path, 'robustness')
-        os.system('mkdir ' + robustness_result_path)
+        os.makedirs(robustness_result_path, exist_ok=True)
 
         print("\n#######################################")
         print("1) Trivial Cases!!")
@@ -1802,7 +979,7 @@ class LLMAdapter:
 
         # Create directory to store results
         trivial_result_path = os.path.join(robustness_result_path, 'trivial')
-        os.system('mkdir ' + trivial_result_path)
+        os.makedirs(trivial_result_path, exist_ok=True)
 
         # # Get Best Prompts
         best_fs_prompt = best_prompts_loaded[model]['FS']
@@ -1829,7 +1006,7 @@ class LLMAdapter:
 
         # Create directory to store results
         non_trivial_result_path = os.path.join(robustness_result_path, 'non-trivial')
-        os.system('mkdir ' + non_trivial_result_path)
+        os.makedirs(non_trivial_result_path, exist_ok=True)
 
         for p, s in prompts:
             self.run_non_trivial_robustness_experiments(
@@ -1852,7 +1029,7 @@ class LLMAdapter:
 
         # Create directory to store results
         real_world_result_path = os.path.join(result_path, 'real-world')
-        os.system('mkdir ' + real_world_result_path)
+        os.makedirs(real_world_result_path, exist_ok=True)
 
         # Run on real world dataset
         scenarios = ['ZT', 'ZR', 'FT', 'FR']
